requirements.txt
ffmpeg==1.4
gTTS==2.3.1
mtranslate==1.8
numpy==1.23.4
openai==0.27.2
Pillow==10.0.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pydub==0.25.1
pyFirmata==1.1.0
pygame==2.3.0
pytesseract==0.3.10
speaker_recognition==0.2
SpeechRecognition==3.10.0
tqdm==4.64.0
